+++ docs/main/changelog.rst
@@ -7,6 +7,12 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+3.18.1
+------
+- Add interpolation of environment variables to ``Configuration.from_yaml()`` and
+  ``Configuration.from_ini()``.
+- Add ignoring of ``IOError`` to ``Configuration.from_yaml()``.
+
 3.18.0
 ------
 - Add ``Configuration.from_yaml()`` method to load configuration from the yaml file.